requirements.txt
torch==2.0.0+cu118
torchvision==0.15.1+cu118
torchaudio==2.0.1+cu118

transformers==4.40.2
accelerate==1.0.1
diffusers==0.31.0
datasets==2.19.1
huggingface-hub==0.26.1
tokenizers==0.19.1
safetensors==0.4.5
peft==0.7.0
ftfy==6.2.0

clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33

albumentations==1.4.6
opencv-python-headless==4.9.0.80
imageio==2.34.1
Pillow>=9.5.0
webdataset==0.2.86
img2dataset==1.45.0
blobfile==3.0.0

numpy==1.24.4
scipy==1.10.1
pandas==2.0.3
scikit-learn==1.3.2
matplotlib==3.7.5
tqdm==4.66.2
pyarrow==15.0.2
PyYAML==6.0.1
rich==11.2.0
fire==0.5.0
psutil==5.9.8
